Added support for compilation databases
So far, this is getting me nowhere but it appears as though the
database is at least parsed correctly right now.

This feels very clunky but of course it makes sense---given a
complicated language such as C++, how should a simple program
know what AST to build without compilation commands?